+++ src/logic/login/PacketSender.h
@@ -16,7 +16,7 @@ public:
     virtual bool Launched(IKernel * kernel);
     virtual bool Destroy(IKernel * kernel);
 
-	virtual void Send(const s32 gate, const s32 actorId, const s32 msgId, const OBuffer& buf, s8 delay = 0);
+	virtual void Send(const s32 gate, const s64 actorId, const s32 msgId, const OBuffer& buf, s8 delay = 0);
 	virtual void Brocast(const std::unordered_map<s32, std::vector<s64>>& actors, const s32 msgId, const OBuffer& buf, s8 delay = 0);
 	virtual void Brocast(const s32 msgId, const OBuffer& buf, s8 delay = 0);
